Vulcano coin source code, which is not very different from Peercoin / novacoin projects, uses the same formula for calculating PoS reward: coin-ages multiplied by yearly percent divided by 365 (approx). Peercoin has 1% interest rate, while Vulcano has 950%, which is clearly indicated in the code. However, the code doesn’t have intended effect, which we have investigated and summarized our investigation in this short report.
A difference between Peercoin and Vulcano is that Peercoin has min stake age of 30 days, which means that a coin used to create a PoS block will not create another block for at least 30 days. Vulcano, on the other hand, has min stake age of 4 hours, which is less than reward base calculation period (1 day). In my opinion, this leads to more quick compounding of coin assets that is desired. It seems that Vulcano authors did not consider the effects of such short stake period on the security of the network. The longer the stake period, the less probable are the attacks of “nothing-at-stake” type, to which these early PoS protocols are vulnerable. That is another reason why we will increase staking period to 30 days.

Also - the higher the interest rate and the higher the frequency of coinstake, the more effective annual percentage is different from the annual percentage.
For 1% annual interest rate and for 30 days min coin stake age effective annual percentage will be just 1.03...
For 100 % interest rate with the same staking frequency it will be 161% etc...
Exact math to correct reward if very difficult, so we wrote simple heuristic. But the higher frequency, the more the actual rewards will diverge from the projected reward percentages.

A task of decreasing PoS reward over time could be done easily. We will hard-code a block number, starting from which nodes will calculate reward by a new formula. We will also update the nodes so that they won’t accept blocks minted with the old formula starting on approximately 01.06.2018. Further decrease of PoS reward will be hard-coded by the number of blocks created, judging by the current frequency of 1 block each approximately 20 seconds.

We decided to make reward calculation depend on the time since the last time user produced a PoS block. This will make yearly interest rate more predictable. This will be implemented as node update, so that new nodes will stop accepting blocks made by old nodes (with high PoS rates) starting on 01.06.2018. Also min stake age will be raised to 30 days. Not only will higher frequency of staking will get less predictable effective annual percentage, but also it will bring back the security heuristics systems of that PoS generation employ to fight against nothing at stake attacks.

Short changelog:

1) The maximum frequency of PoS block generation was reduced to 1 per month.
2) The annual interest was lowered.
3) The formula for calculating the award now looks like :
CoinAge*((AnnualPrecent + 1) ^ (1/frequency) - 1)*frequency*(1/365)

The greater the frequency of payments, the less accurately this formula will work.
4) Prior to block 855,000, the award will be accrued with the desired annual percentage of 400. Between 855,000 and 1,255,000, it will be 300%. Further - from 1,255,000 to 2,610,000 - 150%, then - 75%.
5) Until 0 hours 0 minutes GMT on June 1, the blocks generated by the old client will be accepted. After that - only new. Accordingly, you will need a final upgrade.